fix: rename file to avoid naming duplicates in folders
