+++ practice/src/js/services/product.js
@@ -1,4 +1,4 @@
-import BaseService from './baseService';
+import BaseService from './base';
 import { API } from '../constants/apiUrl';
 
 export default class ProductService extends BaseService {
